Use decimal for money values

+++ db/src/main/resources/db/migration/V3__create_account_table.sql
@@ -2,5 +2,5 @@ create table account(
   id serial primary key,
   owner bigint not null references `user`(id),
   name text not null,
-  initial_balance double precision not null default 0
+  initial_balance decimal(13, 4) not null default 0
 );
